parking-service/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
pytesseract==0.3.10
numpy==1.26.3
pillow==10.1.0